requirements.txt
weaviate-client
streamlit
google
google-generativeai
pillow
requests